fix compile error.
### MR描述
<!--- 详细描述MR的细节 -->

修复编译错误：
EXPECT_EQ宏进行size_t和int之间比较时，编译器会报错。


### 修改动机和上下文背景
<!--- 为什么需要此修改, 解决了什么问题 -->
<!---如果解决了相关的#issue, 在此处进行关联(#issue, close #issue) -->



### 此MR如何进行测试 ?
<!--- 请描述测试MR的细节 -->
<!--- 包括测试的环境以及执行的测试用例 -->
<!--- 说明 change 如何影响其他部分的代码 etc. -->

### change 类型
<!---你的代码引入了何种类型的change, 在所有关联的复选框前选择"x" -->
- [ ] Bug fix (修复了issue的非侵入式修改)
- [ ] New feature (增加功能的非侵入式修改)
- [ ] Breaking change (修复或者增加特性, 但是会造成现有行为的非预期行为)

### 清单
<!--- 查看下述选项,并进行"x"勾选 -->
<!--- 如果你对所有都不确定, 请随时咨询我们 -->
- [ ] 遵循项目的Code-Style
- [ ] Change 需要文档的修改
- [ ] 我已经进行相关文档的修改
- [ ] 我的MR已经通过的相关流水线测试

+++ src/tendisplus/utils/status_test.cpp
@@ -19,8 +19,8 @@ TEST(Status, Common) {
 }
 
 
-size_t callCnt;
-size_t moveCallCnt;
+int32_t callCnt;
+int32_t moveCallCnt;
 class Base {
  public:
   Base() {
